refactor: ProductPageViewContoller 리팩토링 #5
